refactor(combo): extract pure predicates to combo/comboPredicates.ts
