+++ config/quality/file-size-baseline.json
@@ -65,7 +65,7 @@
     "open-sse/services/batchProcessor.ts": 828,
     "open-sse/services/browserBackedChat.ts": 850,
     "open-sse/services/claudeCodeCompatible.ts": 1202,
-    "open-sse/services/combo.ts": 4872,
+    "open-sse/services/combo.ts": 4740,
     "open-sse/services/rateLimitManager.ts": 1022,
     "open-sse/services/tokenRefresh.ts": 1997,
     "open-sse/services/usage.ts": 3408,
